rearrange files in module_symmetry/test
